puts half

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -7,7 +7,7 @@
  */
 void puts_half(char *str)
 {
-	int count = 0;
+	long int count = 0;
 
 	while (str[count] != '\0')
 	{
